refactor: move enums file location
